renamed visual post processor to postprocessor
