+++ clayrs/content_analyzer/information_processor/postprocessors/__init__.py
@@ -1,0 +1,1 @@
+from .postprocessor import *
