--- clayrs/content_analyzer/information_processor/visual_postprocessors/__init__.py
@@ -1,1 +1,0 @@
-from .visualpostprocessor import *
